requirements.txt
pydub==0.25.1
aiogram==3.20.0.post0
pytubefix==8.13.1
yt-dlp==2025.4.30
ffmpeg==1.4
